requirements.txt
mechanicalsoup
pandas
python-dotenv